--- src/contracts/config.ts
@@ -1,11 +1,0 @@
-/**
- * Configuration interface
- */
-export interface ITrConfig {
-  invalidClass?: string;
-  validClass?: string;
-  local?: {
-    lang: string;
-  };
-  realTime?: boolean;
-}
