--- src/contracts/input-change-event.ts
@@ -1,17 +1,0 @@
-import { Rule } from "./rule";
-
-/**
- * tr.input.validated event details
- */
-export type InputEventDetails = {
-  rules: Rule[];
-  element: HTMLInputElement | HTMLTextAreaElement;
-  input: Record<string, string>;
-};
-
-/**
- * Input change event
- */
-export interface InputChangeEvent {
-  details: InputEventDetails;
-}
