--- src/contracts/rule-callback.ts
@@ -1,8 +1,0 @@
-import { InputType, ValidationState } from ".";
-
-/**
- * Rule callback
- */
-export interface RuleCallBack {
-  (input: any, param?: any, type?: InputType): ValidationState;
-}
