--- src/contracts/rules-bag.interface.ts
@@ -1,10 +1,0 @@
-import { Rule } from ".";
-import { RuleCallBack } from "./rule-callback";
-
-export type RulesBag = {
-  [ruleName in Rule]: RuleCallBack;
-};
-
-export type RulesMessages = {
-  [key: string]: string;
-};
