+++ src/contracts/types.ts
@@ -1,4 +1,3 @@
-import { RulesMessages } from ".";
 import { TrivuleForm } from "../validation";
 import { Rule } from "./rule";
